requirements.txt
comfy==0.0.1
diffusers==0.35.1
einops==0.8.1
gradio==5.44.0
huggingface_hub==0.34.4
numpy==2.3.2
onnxruntime==1.22.1
onnxruntime_gpu==1.22.0
optimum==1.27.0
Pillow==11.3.0
PyQt5==5.15.11
safetensors==0.6.2
torch==2.5.1+cu121
torchvision==0.20.1+cu121
transformers==4.53.3